fix: move drawer to the front of the TabBar
