+++ app/src/main/java/to/bitkit/ui/screens/wallets/HomeScreen.kt
@@ -5,10 +5,13 @@ import android.os.Build
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.slideInHorizontally
+import androidx.compose.animation.slideOutHorizontally
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.gestures.snapping.SnapPosition
 import androidx.compose.foundation.gestures.snapping.rememberSnapFlingBehavior
+import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -56,6 +59,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.zIndex
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavController
@@ -132,6 +136,8 @@ fun HomeScreen(
     val homeViewModel: HomeViewModel = hiltViewModel()
     val homeUiState: HomeUiState by homeViewModel.uiState.collectAsStateWithLifecycle()
     val drawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
+    val hasSeenWidgetsIntro by settingsViewModel.hasSeenWidgetsIntro.collectAsStateWithLifecycle()
+    val scope = rememberCoroutineScope()
 
     Box(modifier = Modifier.fillMaxSize()) {
         val walletNavController = rememberNavController()
@@ -145,7 +151,6 @@ fun HomeScreen(
                 val hasSeenShopIntro by settingsViewModel.hasSeenShopIntro.collectAsStateWithLifecycle()
                 val hasSeenProfileIntro by settingsViewModel.hasSeenProfileIntro.collectAsStateWithLifecycle()
                 val quickPayIntroSeen by settingsViewModel.quickPayIntroSeen.collectAsStateWithLifecycle()
-                val hasSeenWidgetsIntro by settingsViewModel.hasSeenWidgetsIntro.collectAsStateWithLifecycle()
 
                 HomeContentView(
                     mainUiState = uiState,
@@ -330,6 +335,149 @@ fun HomeScreen(
                 .align(Alignment.BottomCenter)
                 .systemBarsPadding()
         )
+
+
+        // Drawer overlay and content - moved from AppScaffold to here
+        // Semi-transparent overlay when drawer is open
+        AnimatedVisibility(
+            visible = drawerState.currentValue == DrawerValue.Open,
+            modifier = Modifier
+                .fillMaxSize()
+                .zIndex(10f) // Higher z-index than TabBar
+        ) {
+            Box(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .background(Colors.Black50)
+                    .clickable(
+                        interactionSource = remember { MutableInteractionSource() },
+                        indication = null
+                    ) {
+                        scope.launch {
+                            drawerState.close()
+                        }
+                    }
+            )
+        }
+
+        // Right-side drawer content
+        AnimatedVisibility(
+            visible = drawerState.currentValue == DrawerValue.Open,
+            enter = slideInHorizontally(
+                initialOffsetX = { it }
+            ),
+            exit = slideOutHorizontally(
+                targetOffsetX = { it }
+            ),
+            modifier = Modifier
+                .align(Alignment.TopEnd)
+                .fillMaxHeight()
+                .zIndex(11f) // Higher z-index than overlay
+        ) {
+            DrawerContent(
+                walletNavController = walletNavController,
+                rootNavController = rootNavController,
+                drawerState = drawerState,
+                onClickAddWidget = {
+                    if (!hasSeenWidgetsIntro) {
+                        rootNavController.navigate(Routes.WidgetsIntro)
+                    } else {
+                        rootNavController.navigate(Routes.AddWidget)
+                    }
+                }
+            )
+        }
+    }
+}
+
+
+@Composable
+private fun DrawerContent(
+    walletNavController: NavController,
+    rootNavController: NavController,
+    drawerState: DrawerState,
+    onClickAddWidget: () -> Unit
+) {
+    val scope = rememberCoroutineScope()
+    val drawerWidth = 200.dp
+
+    Column(
+        modifier = Modifier
+            .width(drawerWidth)
+            .fillMaxHeight()
+            .background(Colors.Brand)
+    ) {
+        VerticalSpacer(60.dp)
+
+        DrawerItem(
+            label = stringResource(R.string.wallet__drawer__wallet),
+            iconRes = R.drawable.ic_coins,
+            modifier = Modifier.clickable {
+                scope.launch { drawerState.close() }
+            },
+        )
+
+        HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
+
+        DrawerItem(
+            label = stringResource(R.string.wallet__drawer__activity),
+            iconRes = R.drawable.ic_heartbeat,
+            modifier = Modifier.clickable {
+                walletNavController.navigate(HomeRoutes.AllActivity)
+                scope.launch { drawerState.close() }
+            },
+        )
+
+        HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
+
+        DrawerItem(
+            label = stringResource(R.string.wallet__drawer__contacts),
+            iconRes = R.drawable.ic_users // TODO IMPLEMENT CONTACTS
+        )
+
+        HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
+
+        DrawerItem(
+            label = stringResource(R.string.wallet__drawer__profile),
+            iconRes = R.drawable.ic_user_square, // TODO IMPLEMENT
+        )
+
+        HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
+
+        DrawerItem(
+            label = stringResource(R.string.wallet__drawer__widgets),
+            iconRes = R.drawable.ic_stack,
+            modifier = Modifier.clickable {
+                onClickAddWidget()
+                scope.launch { drawerState.close() }
+            },
+        )
+
+        HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
+
+        DrawerItem(
+            label = stringResource(R.string.wallet__drawer__shop),
+            iconRes = R.drawable.ic_store_front,
+            modifier = Modifier.clickable {
+                rootNavController.navigate(Routes.ShopDiscover)
+                scope.launch { drawerState.close() }
+            },
+        )
+
+        HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
+
+        DrawerItem(
+            label = stringResource(R.string.wallet__drawer__settings),
+            iconRes = R.drawable.ic_settings,
+            modifier = Modifier.clickable {
+                rootNavController.navigateToSettings()
+                scope.launch { drawerState.close() }
+            },
+        )
+
+        HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
+
+        //TODO NAVIGATE TO APP STATE
     }
 }
 
@@ -355,7 +503,6 @@ private fun HomeContentView(
 
     AppScaffold(
         titleText = stringResource(R.string.slashtags__your_name_capital),
-        drawerState = drawerState,
         actions = {
             IconButton(onClick = {
                 scope.launch {
@@ -368,84 +515,6 @@ private fun HomeContentView(
                 )
             }
         },
-        drawerContent = {
-            val drawerWidth = 200.dp
-
-            Column(
-                modifier = Modifier
-                    .width(drawerWidth)
-                    .fillMaxHeight()
-                    .background(Colors.Brand)
-            ) {
-                VerticalSpacer(60.dp)
-
-                DrawerItem(
-                    label = stringResource(R.string.wallet__drawer__wallet),
-                    iconRes = R.drawable.ic_coins,
-                    modifier = Modifier.clickable {
-                        scope.launch { drawerState.close() }
-                    },
-                )
-
-                HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
-
-                DrawerItem(
-                    label = stringResource(R.string.wallet__drawer__activity),
-                    iconRes = R.drawable.ic_heartbeat,
-                    modifier = Modifier.clickable {
-                        walletNavController.navigate(HomeRoutes.AllActivity)
-                    },
-                )
-
-                HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
-
-                DrawerItem(
-                    label = stringResource(R.string.wallet__drawer__contacts),
-                    iconRes = R.drawable.ic_users // TODO IMPLEMENT CONTACTS
-                )
-
-                HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
-
-                DrawerItem(
-                    label = stringResource(R.string.wallet__drawer__profile),
-                    iconRes = R.drawable.ic_user_square, // TODO IMPLEMENT
-                )
-
-                HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
-
-                DrawerItem(
-                    label = stringResource(R.string.wallet__drawer__widgets),
-                    iconRes = R.drawable.ic_stack,
-                    modifier = Modifier.clickable {
-                        onClickAddWidget()
-                    },
-                )
-
-                HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
-
-                DrawerItem(
-                    label = stringResource(R.string.wallet__drawer__shop),
-                    iconRes = R.drawable.ic_store_front,
-                    modifier = Modifier.clickable {
-                        rootNavController.navigate(Routes.ShopDiscover)
-                    },
-                )
-
-                HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
-
-                DrawerItem(
-                    label = stringResource(R.string.wallet__drawer__settings),
-                    iconRes = R.drawable.ic_settings,
-                    modifier = Modifier.clickable {
-                        rootNavController.navigateToSettings()
-                    },
-                )
-
-                HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
-
-                //TODO NAVIGATE TO APP STATE
-            }
-        }
     ) {
         RequestNotificationPermissions()
         val balances = LocalBalances.current
